+++ Services/Remote/RemoteHostTelemetryEventArgs.cs
@@ -4,6 +4,7 @@ namespace NetKeyer.Services.Remote;
 
 public class RemoteHostTelemetryEventArgs : EventArgs
 {
+    public bool IsTransmitModeCW { get; init; } = true;
     public double LastLagMs { get; init; }
     public double AvgLagMs { get; init; }
     public double MaxLagMs { get; init; }